app/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
PyYAML
